requirements.txt
Django
djangorestframework==3.14.0
pymongo==4.6.1
